+++ demo/.eslintrc.js
@@ -6,7 +6,7 @@ module.exports = {
   extends: ['plugin:react/recommended'],
   settings: {
     react: {
-      version: 'latest',
+      version: 'detect',
     },
   },
   rules: {