+++ CHANGELOG.md
@@ -3,8 +3,8 @@
 ## Unreleased
 
 - [maintenance] Upgrade all dependencies
-- [fix] Upgrade Playwright from 1.16.3 to 1.18.0
-- [fix] Upgrade Yargs from 17.2.1 to 17.3.1
+- [fix] Upgrade Playwright from 1.16.3 to 1.21.1
+- [fix] Upgrade Yargs from 17.2.1 to 17.4.1
 
 ## 5.0.0 (2021-11-29)